split lime into lime text and lime image
